Get rid of `WorkflowRunner` entirely by collapsing into Triton model
